third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-lifetime-expected.txt

Tests the RTCPeerConnection lifetime.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS observationA.wasCollected is true
PASS observationB.wasCollected is false
PASS dataChannelOpen
PASS dc.send('xyzzy'); did not throw exception.
PASS dataChannelMessage
PASS successfullyParsed is true

TEST COMPLETE
